+++ ddsrouter_core/src/cpp/reader/implementations/rtps/Reader.cpp
@@ -22,6 +22,7 @@
 #include <reader/implementations/rtps/Reader.hpp>
 #include <ddsrouter_utils/exception/InitializationException.hpp>
 #include <ddsrouter_utils/Log.hpp>
+#include <ddsrouter_utils/format/Formatter.hpp>
 
 namespace eprosima {
 namespace ddsrouter {
