+++ ddsrouter_utils/include/ddsrouter_utils/event/impl/EventHandler.ipp
@@ -23,6 +23,7 @@
 
 #include <ddsrouter_utils/utils.hpp>
 #include <ddsrouter_utils/Log.hpp>
+#include <ddsrouter_utils/format/Formatter.hpp>
 
 namespace eprosima {
 namespace ddsrouter {
@@ -33,8 +34,7 @@ const std::function<void(Args...)> EventHandler<Args...>::DEFAULT_CALLBACK_ =
         [](Args...)
         {
             // This should never happen
-            utils::tsnh(
-                utils::Formatter() << "This callback must not be called.");
+            utils::tsnh("This callback must not be called.");
         };
 
 template <typename ... Args>
