+++ ddsrouter_utils/test/unittest/memory/CMakeLists.txt
@@ -21,7 +21,7 @@ set(TEST_NAME OwnerPtrTest)
 set(TEST_SOURCES
         OwnerPtrTest.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/exception/Exception.cpp
-        ${PROJECT_SOURCE_DIR}/src/cpp/Formatter.cpp
+        ${PROJECT_SOURCE_DIR}/src/cpp/format/Formatter.cpp
     )
 
 set(TEST_LIST
@@ -54,7 +54,7 @@ set(TEST_NAME LesseePtrTest)
 set(TEST_SOURCES
         LesseePtrTest.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/exception/Exception.cpp
-        ${PROJECT_SOURCE_DIR}/src/cpp/Formatter.cpp
+        ${PROJECT_SOURCE_DIR}/src/cpp/format/Formatter.cpp
     )
 
 set(TEST_LIST
