+++ ddsrouter_utils/test/unittest/thread_pool/CMakeLists.txt
@@ -28,7 +28,7 @@ set(TEST_SOURCES
         ${PROJECT_SOURCE_DIR}/src/cpp/wait/IntWaitHandler.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/wait/CounterWaitHandler.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/utils.cpp
-        ${PROJECT_SOURCE_DIR}/src/cpp/Formatter.cpp
+        ${PROJECT_SOURCE_DIR}/src/cpp/format/Formatter.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/exception/Exception.cpp
     )
