+++ ddsrouter_utils/test/unittest/utils/CMakeLists.txt
@@ -17,7 +17,8 @@ set(TEST_NAME utilsTest)
 set(TEST_SOURCES
         utilsTest.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/exception/Exception.cpp
-        ${PROJECT_SOURCE_DIR}/src/cpp/Formatter.cpp
+        ${PROJECT_SOURCE_DIR}/src/cpp/format/Formatter.cpp
+        ${PROJECT_SOURCE_DIR}/src/cpp/format/format_utils.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/math/math.cpp
         ${PROJECT_SOURCE_DIR}/src/cpp/utils.cpp
     )
